Fix model definitions for SQLite compatibility

+++ models/agent_supplier.js
@@ -18,7 +18,7 @@ Adress:{
     allowNull:false
 },
 Phone:{
-    type:Sequelize.BIGINT(20),
+    type:Sequelize.INTEGER,
     allowNull:false,
     },
 Email:{
